+++ src/components/install/DesktopSettingsConfiguration.vue
