+++ src/components/install/InstallLocationPicker.vue
