+++ src/components/install/MigrationPicker.vue
