+++ src/router.ts
@@ -1,12 +1,25 @@
 import {
   createRouter,
   createWebHashHistory,
-  createWebHistory
+  createWebHistory,
+  NavigationGuardNext,
+  RouteLocationNormalized
 } from 'vue-router'
 import LayoutDefault from '@/views/layouts/LayoutDefault.vue'
 import { isElectron } from './utils/envUtil'
 
 const isFileProtocol = () => window.location.protocol === 'file:'
+const guardElectronAccess = (
+  to: RouteLocationNormalized,
+  from: RouteLocationNormalized,
+  next: NavigationGuardNext
+) => {
+  if (isElectron()) {
+    next()
+  } else {
+    next('/')
+  }
+}
 
 const router = createRouter({
   history: isFileProtocol() ? createWebHashHistory() : createWebHistory(),
@@ -24,27 +37,19 @@ const router = createRouter({
           path: 'server-start',
           name: 'ServerStartView',
           component: () => import('@/views/ServerStartView.vue'),
-          beforeEnter: async (to, from, next) => {
-            // Only allow access to this page in electron environment
-            if (isElectron()) {
-              next()
-            } else {
-              next('/')
-            }
-          }
+          beforeEnter: guardElectronAccess
         },
         {
           path: 'install',
           name: 'InstallView',
           component: () => import('@/views/InstallView.vue'),
-          beforeEnter: async (to, from, next) => {
-            // Only allow access to this page in electron environment
-            if (isElectron()) {
-              next()
-            } else {
-              next('/')
-            }
-          }
+          beforeEnter: guardElectronAccess
+        },
+        {
+          path: 'welcome',
+          name: 'WelcomeView',
+          component: () => import('@/views/WelcomeView.vue'),
+          beforeEnter: guardElectronAccess
         }
       ]
     }
