+++ src/views/WelcomeView.vue
@@ -1,29 +1,33 @@
 <template>
-  <div class="flex flex-col items-center justify-center gap-8 p-8">
-    <!-- Header -->
-    <h1 class="animated-gradient-text text-glow">Welcome to ComfyUI</h1>
+  <div
+    class="font-sans flex flex-col justify-center items-center h-screen m-0 text-neutral-300 bg-neutral-900 dark-theme pointer-events-auto"
+  >
+    <div class="flex flex-col items-center justify-center gap-8 p-8">
+      <!-- Header -->
+      <h1 class="animated-gradient-text text-glow">Welcome to ComfyUI</h1>
 
-    <!-- Get Started Button -->
-    <Button
-      label="Get Started"
-      icon="pi pi-arrow-right"
-      iconPos="right"
-      size="large"
-      rounded
-      @click="$emit('start')"
-      class="p-4 text-lg fade-in-up"
-    />
+      <!-- Get Started Button -->
+      <Button
+        label="Get Started"
+        icon="pi pi-arrow-right"
+        iconPos="right"
+        size="large"
+        rounded
+        @click="$router.push('/install')"
+        class="p-4 text-lg fade-in-up"
+      />
+    </div>
   </div>
 </template>
 
 <script setup lang="ts">
 import Button from 'primevue/button'
-defineEmits(['start'])
 </script>
 
 <style scoped>
 .animated-gradient-text {
-  @apply font-bold text-[clamp(2rem,8vw,4rem)];
+  @apply font-bold;
+  font-size: clamp(2rem, 8vw, 4rem);
   background: linear-gradient(to right, #12c2e9, #c471ed, #f64f59, #12c2e9);
   background-size: 300% auto;
   background-clip: text;
@@ -40,6 +44,7 @@ defineEmits(['start'])
   0% {
     background-position: 0% center;
   }
+
   100% {
     background-position: 300% center;
   }
@@ -55,6 +60,7 @@ defineEmits(['start'])
     opacity: 0;
     transform: translateY(20px);
   }
+
   100% {
     opacity: 1;
     transform: translateY(0);
